The fixture create_psd_from_txt_data() (which is duplicated in several modules) was moved in the conftest.py module.
