+++ tests/conftest.py
@@ -23,6 +23,8 @@
 import os
 
 from ansys.dpf.core import upload_file_in_tmp_folder
+from ansys.dpf.core import TimeFreqSupport, fields_factory, locations
+import numpy as np
 import pytest
 
 from ansys.sound.core.server_helpers import connect_to_or_start_server
@@ -245,3 +247,50 @@ def pytest_configure(config):
     else:
         # Local server => we use the repository folder structure where the output folder exists.
         pytest.output_folder = os.path.join(os.path.dirname(pytest.data_path_flute), "output")
+
+
+@pytest.fixture
+def create_psd_from_txt_data():
+    """Create a PSD DPF field from the flute_psd.txt test data file."""
+    path_flute_psd = pytest.data_path_flute_psd_locally
+
+    # Open a txt file for reading
+    fid = open(path_flute_psd)
+
+    # skip first line
+    fid.readline()
+
+    # read all other lines
+    all_lines = fid.readlines()
+    # close file
+    fid.close()
+
+    amplitudes = []
+
+    for line in all_lines:
+        splitted_line = line.split()
+        amplitudes.append(float(splitted_line[1]))
+
+    amplitudes = np.array(amplitudes)
+
+    # convert dBSPL / Hz -> Pa^2/Hz
+    amplitudes = np.power(10, amplitudes / 10)
+    amplitudes = amplitudes * 2.0e-5
+    amplitudes = amplitudes * 2.0e-5
+
+    # for now, due to a sharp constraint on regularity of frequencies (1.0e-5) in the operator,
+    # we cannot use those written in the file. Let's recreate them
+    frequencies = np.linspace(0, 22050, len(amplitudes))
+
+    psd = fields_factory.create_scalar_field(num_entities=1, location=locations.time_freq)
+    psd.append(amplitudes, 1)
+    support = TimeFreqSupport()
+    frequencies_field = fields_factory.create_scalar_field(
+        num_entities=1, location=locations.time_freq
+    )
+    frequencies_field.append(frequencies, 1)
+    support.time_frequencies = frequencies_field
+
+    psd.time_freq_support = support
+
+    yield psd
